Gqbranch

+++ lib/src/utils/database_helper.dart
@@ -40,7 +40,6 @@ class DatabaseHelper {
 // meal table:
   String mealTable = 'meal_table';
   String colId = 'mealId'; //auto
-  //String _email;
   String colMealSlot = 'slot';
   String colCarb = 'totalCarb';
   String colnote = 'note';
@@ -48,37 +47,25 @@ class DatabaseHelper {
 
 // variety table:
   String varietyTable = 'variety_table';
-  String colid = 'varId';
-  //int _mealId;
-  //String _email;
+  String colid = 'id';
   String coleat = 'eat';
   String colvarcarb = 'varcarb';
   String colamount = 'amount';
   
   // BG table:
-  String BGTable= 'BG_table';
-  //colEmail
+  String bGTable= 'BG_table';
   String colBGSlot = 'slot';
   String colBg='BG';
- // colnote
-// String colDg='dg';
 
-   // BP table:
-  String BPTable= 'BP_table';
-  //colEmail
+
+// BP table:
+  String bPTable= 'BP_table';
   String colDiastolic = 'Diastolic';
   String colSystolic='Systolic';
- // colnote
-// String colDp='dp';
-
 
 //BW table:
-  String BWTable= 'BW_table';
- // String _email;
+  String bWTable= 'BW_table';
   String colwieght='wieght';
- // double _bmi;
- // String _note;
- // String _dw;
 
 //Instruction table:
   String instrTable= 'Instr_table';
@@ -87,60 +74,37 @@ class DatabaseHelper {
   String coldes='des';
 
 //PA table:
-  String PATable= 'PA_table';
-    // this._email = map['email'];
+  String pATable= 'PA_table';
   String colName = 'Name';
-  //String colPAid='id'; //auto
   String coldur = 'dur';
-  //  this._dg = map['date'];
 
 //PT table:
   String PTTable= 'PT_table';
-//  String colName = 'Name';
-  //String colPTid='id'; //auto
 
 //A1C table:
-  String A1CTable= 'A1C_table';
-//  String colName = 'Name';
- // String colid='id'; //auto
+  String a1CTable= 'A1C_table';
   String colA1C='a1C';
   String coldS='dateS';
   String coldE='dateE';
 
 //CARB table:
-  String CarbTable= 'Carb_table';
-  //String colName = 'Name';
-  //String colA1Cid='id'; //auto
+  String carbTable= 'Carb_table';
   String colcurnt='curnt';
   String colmin='min';
   String colmax='max';
-  //String _Dc;
 
 //Exams table:
   String ExamsTable= 'Exams_table';
-  //String _id;//auto
   String colPTId='PTId';
-  //String _email;
-  //String _name;
-  //String _dur;
-  //String _date;
   String colresult='result';
   String colRDate='RDate';
 
 //ExamA1C table:
   String ExamA1CTable= 'ExamA1C_table';
-  //String _id;//auto
-  //String _PTId='PTId';
-  //String _email;
-  //String _name;
-  //String _dur;
-  //String _date;
-  //String _result='result';
-  //String _RDate='RDate';
+
 
 //med table:
-  //int _id; //auto
-  //String _Name;
+
   String MedTable= 'med_table';
   String colconf='conf';
   String colwork='work';
@@ -149,10 +113,7 @@ class DatabaseHelper {
 
 //dug table:
   String DugTable= 'Dug_table';
-  //int _id; //auto
-  //String _email;
-  //String _Name;
-  //String _type;
+
   String coldug='dug';
 
 
@@ -171,7 +132,7 @@ class DatabaseHelper {
 
   Future<Database> initializeDatabase() async {
     Directory directory = await getApplicationDocumentsDirectory();
-    String path = directory.path + 'user7.db';
+    String path = directory.path + 'user11.db';
     var userDatabase =
         await openDatabase(path, version: 1, onCreate: _createDb);
     return userDatabase;
@@ -191,18 +152,18 @@ class DatabaseHelper {
 
     await db.execute('CREATE TABLE $varietyTable ('
         '$colid INT,$colId INT,$coleat TEXT,'
-        'colEmail TEXT, $colvarcarb REAL,'
+        '$colEmail TEXT, $colvarcarb REAL,'
         '$colamount REAL,PRIMARY KEY ($colid,$colId))');
 
-    await db.execute('CREATE TABLE $BGTable ('
+    await db.execute('CREATE TABLE $bGTable ('
         '$colEmail TEXT,$coldate TEXT,$colBGSlot TEXT,'
         ' $colBg INT,$colnote TEXT,PRIMARY KEY ($colEmail,$coldate))');
     
-    await db.execute('CREATE TABLE $BPTable ('
+    await db.execute('CREATE TABLE $bPTable ('
         '$colEmail TEXT,$coldate TEXT,$colSystolic INT,'
         ' $colDiastolic INT,$colnote TEXT,PRIMARY KEY ($colEmail,$coldate))');
 
-     await db.execute('CREATE TABLE $BWTable ('
+     await db.execute('CREATE TABLE $bWTable ('
         '$colAototId INTEGER PRIMARY KEY AUTOINCREMENT,'
         '$colEmail TEXT,$coldate TEXT,$colwieght INT,'
         ' $colBmi REAL,$colnote TEXT)');
@@ -211,7 +172,7 @@ class DatabaseHelper {
     '$colAototId INTEGER PRIMARY KEY AUTOINCREMENT,'
         '$coltitel TEXT , $coldes TEXT)');
 
-    await db.execute('CREATE TABLE $PATable('
+    await db.execute('CREATE TABLE $pATable('
         '$colAototId INTEGER PRIMARY KEY AUTOINCREMENT,'
         '$colEmail TEXT ,$coldate TEXT ,$colName TEXT , $coldur REAL)');
 
@@ -219,7 +180,7 @@ class DatabaseHelper {
         '$colAototId INTEGER PRIMARY KEY AUTOINCREMENT,'
         '$colName TEXT)');
   
-    await db.execute( 'CREATE TABLE $A1CTable('
+    await db.execute( 'CREATE TABLE $a1CTable('
     '$colAototId INTEGER PRIMARY KEY AUTOINCREMENT,'
         '$colEmail TEXT,$coldS TEXT ,$coldE TEXT ,$colA1C TEXT)');
 
@@ -235,9 +196,9 @@ class DatabaseHelper {
         '$colName TEXT, $coldate TEXT, $colresult TEXT, $colRDate TEXT,'
         '$coldur INTEGER)');
     
-    await db.execute( 'CREATE TABLE $CarbTable('
+    await db.execute( 'CREATE TABLE $carbTable('
     '$colAototId INTEGER PRIMARY KEY AUTOINCREMENT,'
-        '$colcurnt REAL ,$colmin REAL ,$colmax REAL,$coldate TEXT )');
+        '$colEmail TEXT,$colcurnt REAL ,$colmin REAL ,$colmax REAL,$coldate TEXT )');
 
     await db.execute( 'CREATE TABLE $DugTable('
     '$colAototId INTEGER PRIMARY KEY AUTOINCREMENT,'
@@ -274,6 +235,16 @@ class DatabaseHelper {
     Database db = await this.database;
     var result = await db.insert(mealTable, meal.toMap());
   //  print(result);
+    return result;}
+
+// get meal id :
+  Future<List<Map<String, dynamic>>> getNoteList( String email) async {
+  Database db = await this.database;
+    var result =await db.query(mealTable,where:"$colEmail =\"$email\"",orderBy:"$colId DESC",limit: 1);
+    if (result == null) {
+     return null;
+    }
+    print("result from meal :$result");
     return result;
   }
 
@@ -292,15 +263,19 @@ class DatabaseHelper {
     return result;
   }
 
-//get meals //not complite
-  Future<List<Map<String, dynamic>>> getMeal(String email) async {
+//get meals
+  Future  getMeal(String email) async {
     Database db = await this.database;
     var result =
-        await db.query(mealTable, where: '$colEmail = ?', whereArgs: [email]);
-    if (result.length > 0) {
-      return result;
-    }
-    return null;
+        await db.rawQuery('SELECT * FROM $mealTable  WHERE  $colEmail=\"$email\" ORDER BY $colId DESC LIMIT 1 ');
+   print(" db.  meal: ${result.toList()}");
+
+if(result.isEmpty){
+  return [];
+}
+else{
+  print("hi meal id from db");
+ return result;}
   }
   
 //-------------------------------Variety Table--------------------------------------------
@@ -319,12 +294,48 @@ class DatabaseHelper {
   }
 
  //delete one variety
-  Future<int> deleteVariety(int id, int mealid) async {
+  Future<int> deleteVariety(int id,String email, int mealid) async {
   	var db = await this.database;
-  	int result = await db.rawDelete('DELETE FROM $varietyTable WHERE $colid = $id AND $colId=$mealid');
+  	int result = await db.rawDelete('DELETE FROM $varietyTable WHERE $colid = $id AND $colId=$mealid AND $colEmail=\"$email\"');
   	return result;
   }
 
+//get vart by id meal:  
+Future getvr(int meal,) async {
+ Database db = await this.database;
+var result = await db.rawQuery('SELECT * FROM $varietyTable WHERE $colId= $meal');
+ print("getvart: ${result.toList()}");
+
+  print("hi varty list");
+ return result;
+}
+
+//get vart by email: noo need 
+Future getfood(String email,) async {
+ Database db = await this.database;
+        var id=await db.rawQuery('SELECT * FROM $mealTable  WHERE  $colEmail=\"$email\" ORDER BY $colId DESC LIMIT 1 ');
+      print("from get food mesl id = $id");
+      print("from get food mesl id = ${id[0]['$colId']}");
+
+        var result = await db.rawQuery('SELECT * FROM $varietyTable WHERE $colEmail=\"$email\"  AND $colId=${id[0]['$colId']} ');
+ print("getvart list: ${result.toList()}");
+
+ if(result.isEmpty) return [];
+ else 
+ return result;//.toList();
+}
+
+
+// //get vart by email: noo need 
+// Future getfood(String email,) async {
+//  Database db = await this.database;
+// var result = await db.rawQuery('SELECT * FROM $varietyTable WHERE $colEmail=\"$email\" AND $colId=$colId=$meal ');
+//  print("getvart list: ${result.toList()}");
+
+//  if(result.isEmpty) return [];
+//  else 
+//  return result;
+// }
  //delete All varietise meal
   Future<int> deleteAllVarieteis(int id, int mealid) async {
   	var db = await this.database;
@@ -335,14 +346,14 @@ class DatabaseHelper {
 //add
   Future<int> insertBG(BG bg) async {
     Database db = await this.database;
-    var result = await db.insert(BGTable, bg.toMap());
+    var result = await db.insert(bGTable, bg.toMap());
     return result;
   }
 
 //get total BG
 Future BGTotal(String email) async {
-  var dbClient = await this.database;
-  var result = await dbClient.rawQuery("SELECT SUM($colBg) as Total FROM $BGTable  WHERE $colEmail=\"$email\"");
+  var db = await this.database;
+  var result = await db.rawQuery("SELECT SUM($colBg) as Total FROM $bGTable  WHERE $colEmail=\"$email\"");
   //print(result.toList());
   return result.toList();
 }
@@ -350,26 +361,26 @@ Future BGTotal(String email) async {
 // nuber of record:
 Future BGRecord(String email) async {
  Database db = await this.database;
-var result = await db.rawQuery('SELECT COUNT(*) as r FROM $BGTable  WHERE  $colEmail=\"$email\"');
+var result = await db.rawQuery('SELECT COUNT(*) as r FROM $bGTable  WHERE  $colEmail=\"$email\"');
   print(result.toList());
   return result.toList();
 }
 
 // GET BG LIST FOR ONE USER:
-  Future<List<Map<String, dynamic>>> getNoteList( String email) async {
-  Database db = await this.database;
-    var result =await db.query(BGTable,where:"$colEmail =\"$email\"",orderBy:"$coldate DESC"  );
-    if (result == null) {
-     return null;
-    }
-    return result;
-  }
+  // Future<List<Map<String, dynamic>>> getNoteList( String email) async {
+  // Database db = await this.database;
+  //   var result =await db.query(bGTable,where:"$colEmail =\"$email\"",orderBy:"$coldate DESC"  );
+  //   if (result == null) {
+  //    return null;
+  //   }
+  //   return result;
+  // }
 
 //----------------------------------Prusser Table--------------------------------------------------- 
 //add
   Future<int> insertBP(BP bp) async {
     Database db = await this.database;
-    var result = await db.insert(BPTable, bp.toMap());
+    var result = await db.insert(bPTable, bp.toMap());
     return result;
   }
   
@@ -377,16 +388,17 @@ var result = await db.rawQuery('SELECT COUNT(*) as r FROM $BGTable  WHERE  $colE
 //add
   Future<int> insertBW(BW bw) async {
     Database db = await this.database;
-    var result = await db.insert(BWTable, bw.toMap());
+    var result = await db.insert(bWTable, bw.toMap());
     return result;
   }
 
 //get wight
  Future<List<Map<String, dynamic>>> getWight(String email) async {
     Database db = await this.database;
-   var result = await db.rawQuery('SELECT $colwieght as wit,$colBmi as bmi FROM $BWTable  WHERE  $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');  
-print(result.toList());
-    
+   var result = await db.rawQuery('SELECT $colwieght as wit,$colBmi as bmi FROM $bWTable  WHERE  $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');  
+//print(result.toList());
+    //   var result = await db.query(userTable, where: '$colEmail = ?', whereArgs: [email],orderBy: '$colAototId DESC',limit: 1 );  
+
     return result.toList();
   }
 //----------------------------------Instruction Table--------------------------------------------------- 
@@ -413,7 +425,7 @@ Future<List<Map<String, dynamic>>> getInstructionMapList() async {
 //add
   Future<int> insertPA(PA pa) async {
     Database db = await this.database;
-    var result = await db.insert(PATable, pa.toMap());
+    var result = await db.insert(pATable, pa.toMap());
 print('updated: $result');
     return result;
   }
@@ -421,9 +433,9 @@ print('updated: $result');
   Future<int> UpdatetPA(int pa) async {
     Database db = await this.database;
  //   int count = await database.rawUpdate(SELECT * FROM tablename ORDER BY column DESC LIMIT 1);
-    var result = await db.rawQuery('SELECT * FROM $PATable ORDER BY $colAototId DESC LIMIT 1');
-   // var result1 = await db.update(PATable, pa.map(), where: '$colAototId = ?', whereArgs: [variety.id]));
-    var result1 = await db.rawUpdate('UPDATE $PATable  SET $coldur = $pa  WHERE colAototId=${result[0]['id']}');
+    var result = await db.rawQuery('SELECT * FROM $pATable ORDER BY $colAototId DESC LIMIT 1');
+   // var result1 = await db.update(pATable, pa.map(), where: '$colAototId = ?', whereArgs: [variety.id]));
+    var result1 = await db.rawUpdate('UPDATE $pATable  SET $coldur = $pa  WHERE $colAototId=${result[0]['id']}');
 print('updated: $result');
 print('updated 1: $result1');
     return result1;
@@ -444,32 +456,38 @@ print('updated 1: $result1');
 //add
   Future<int> insertA1C(A1C a1c) async {
     Database db = await this.database;
-    var result = await db.insert(A1CTable, a1c.toMap());
+    var result = await db.insert(a1CTable, a1c.toMap());
     return result;
   }
   
    //UPDATE
    Future<int> UpdatetA1C(double a1c,String email) async {
      Database db = await this.database;
-    var result1 = await db.rawQuery('SELECT $colAototId  FROM $A1CTable WHERE  $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');
-    var result = await db.rawUpdate('UPDATE $A1CTable  SET $colA1C = $a1c  WHERE $colAototId=${(result1.toList())[0]['id']}');
+    var result1 = await db.rawQuery('SELECT $colAototId  FROM $a1CTable WHERE  $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');
+    var result = await db.rawUpdate('UPDATE $a1CTable  SET $colA1C = $a1c  WHERE $colAototId=${(result1.toList())[0]['id']}');
   print("UpdatetA1C:$a1c, ( $result1 ),($result)");
     return result;
    }
 
 //GET last value of A1C:
-Future getA1C(String email) async {
+Future<List<Map<String, dynamic>>> getA1C(String email) async {
  Database db = await this.database;
-var result = await db.rawQuery('SELECT $colA1C FROM $A1CTable WHERE $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');
- // print("getA1C: ${result.toList()}");
+var result = await db.rawQuery('SELECT $colA1C FROM $a1CTable WHERE $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');
+ print("getA1C: ${result.isEmpty}");
+result.toList();
 
-  print("hi list");
-  return result;
+
+if(result.isEmpty){ //result.add({'a1C':0.0});
+  return [];}
+  else{
+  print("hi list: $result");
+
+  return result;}
 }
 //GET last value of id:
 Future getA1Cid(String email) async {
  Database db = await this.database;
-var result = await db.rawQuery('SELECT $colAototId FROM $A1CTable WHERE $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');
+var result = await db.rawQuery('SELECT $colAototId FROM $a1CTable WHERE $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');
 print(result.toList());
 
   return result.toList();
@@ -478,7 +496,7 @@ print(result.toList());
 // nuber of record:
 Future A1CRecord(String email) async {
  Database db = await this.database;
-var result = await db.rawQuery('SELECT COUNT(*) as r FROM $A1CTable WHERE $colEmail=\"$email\"');
+var result = await db.rawQuery('SELECT COUNT(*) as r FROM $a1CTable WHERE $colEmail=\"$email\"');
   print("A1CRecord ${result.toList()}");
   return result.toList();
 }
@@ -491,13 +509,31 @@ var result = await db.rawQuery('SELECT COUNT(*) as r FROM $A1CTable WHERE $colEm
 //add
   Future<int> insertCARB(Carb pa) async {
     Database db = await this.database;
-    var result = await db.insert(CarbTable, pa.toMap());
+    var result = await db.insert(carbTable, pa.toMap());
     return result;
   }
 
 //UPDATE
-//GET
+Future<int> updataCarb(String email,double carb,String date) async {
+ Database db = await this.database;
+var result = await db.rawUpdate('UPDATE $carbTable SET $colcurnt =$carb+$colcurnt   WHERE $colEmail = \"$email\" AND $coldate=\"$date\"');
+ print("update Carb: ${result}");
+return result;
+}
 
+//GET last value:
+Future<List<Map<String, dynamic>>> getCarb(String email) async {
+ Database db = await this.database;
+var result = await db.rawQuery('SELECT * FROM $carbTable WHERE $colEmail=\"$email\" ORDER BY $colAototId  DESC LIMIT 1 ');
+ print("getCarb: ${result.isEmpty}");
+ 
+if(result.isEmpty){ //result.add({'a1C':0.0});
+  return [];}
+  else{
+  print("carbe list: $result");
+
+  return result;}
+}
 //----------------------------------Exam Table---------------------------------------------------
 //add
   Future<int> insertExam(Exam pa) async {
@@ -506,12 +542,12 @@ var result = await db.rawQuery('SELECT COUNT(*) as r FROM $A1CTable WHERE $colEm
     return result;
   }
 //UPDATE
+
 //GET
 Future getPT(String email) async {
  Database db = await this.database;
 var result = await db.rawQuery('SELECT * FROM $ExamsTable WHERE $colEmail=\"$email\" ORDER BY $coldur  DESC  ');
  print("getA1C: ${result.toList()}");
-
   print("hi list");
  return result;
 }
